Fix XDG_CURRENT_DESKTOP detection for multiple desktop name.
